chore: ContentCard 폴더 구조 변경
