+++ src/App.jsx
@@ -1,4 +1,11 @@
+import ContentCard from "./components/Card/ContentCard";
+
 function App() {
-    return <h1>hello world</h1>;
+    return <ContentCard 
+        companyName="Company A"
+        deadline="마감 D-2"
+        contentName="Career Content Name"
+        scrap={false}
+    />
 }
 export default App;
